RELENG-365 Create pipelines for ForgeOps Docker builds
The ForgeOps repo contains several Dockerfiles, which only need to be
built if they (or files related to them) have changed since the previous
build.

For postcommit builds, Docker images are build locally and pushed to
GCR, tagged with '7.0.0-{gitCommitSha}' and 'latest'

For PR builds, the Docker images are only build locally. Due to
difficulties comparing the PR branch with the target branch, ALL images
are built on the first run of a PR.

Images/directories can be ignored by commenting them out of the
`buildDirectories` map in the Jenkinsfile.