Analyse file changes using PR-/postcommit-specific methods
Comparing changes between the current build and the previous one is the
best approach for postcommit builds, which tend to happen incrementally
each time a new commit is added to a branch, and have no alternative
branch to compare with.

Comparing changes against a target branch is best suited to PR builds;
commits on these branches can freely be added and removed, or the PR
rebased off the target branch. Squashing and rebasing can cause issues
when comparing file changes between builds.