Updates required for cert-manager 0.8.0
Cert manager now supports an ingress shim. It is not longer required to create a Certificate CRD.
If the ingress is annotated with the name of an issuer, cert-manager will auto
create the certificate request.

This commit removes the cert-manager configs from kustomize. The deploy-certmanager.sh
script has been enhanced to create both the ca-issuer and the Let Encrypt issuer.