refactor: CLOUD-2076 Add enduser path to ingress in 6.5
Added enduser path to 6.5 forgerock ingress and this was missing.
Because of the difference between 6.5 and 7.0 ingresses, the ingress folders have been moved into 6.5 and 7.0 base folders.
The enduser UI and platform UI paths have been removed from the 6.5 forgerock ingress as these are 7.0 only.
Also minor updates for foregops-secrets readme.