feat: Add tekton "prod" pipeline
Adds a tekton pipeline for the the prod deploy. This is similar to the nightly
but it does not install from scratch. The user and application stores are preserved
and AM and IDM are upgraded.

This PR also includes improvements to Tekton Task failure handling. All failed tasks
will send a notification to the #cloud-deploy-notify channel.

The existing pipelines have not been modified to use the failure handler. This wull happen
after we gain some experience with the approach.

refs: CLOUD-2409