feat: CDM container uses stack file in localdirectory
Prior to this change the cdm container would use only the stack file in
the container. This doesn't work well with using custom stack
configurations. Running the cdm container does cause a side effect on
the project file because it's changed to point to the index.ts in the
container. This is required because the container runs with code that's
shipped with the container and not the ts code in the current directory.

jira: CLOUD-2225