RELENG-808 Convert PIT#2 promotion job to a multibranch pipeline
Potentially, there will be multiple branches that need to perform PIT#2
promotions, even if using branches to do this is not the long-term
strategy. For that reason, re-write the PIT#2 pipeline as a multibranch
pipeline, so that each branch can maintain its own version of the
pipeline.

Using a multibranch pipeline allows the early PaaS release branches to
make the customisations they need to the PIT#2 tests and promotion
mechanism, without interfering with the configuration on master.

Only the 'master' branch has a scheduled build; the jobs for all other
branches will need to be launched manually.

As it is unclear what kind of future branches will be required to
perform PIT#2 promotion, hard-code logic into the pipeline that only
allow 'master' to be built. This can be changed in the pipeline code on
any PaaS release branch, once that branch is forked from master. This is
a safeguard to prevent erroneous PIT#2 promotions from illegitimate
branches.