feat: Adds a sidecar for AM and IDM. Provides a slimmed down development instance
Adds sidecars that push changes to a git-server pod that runs in the same namespace.
Also includes updates to the cli-tools to support the git server and the toolbox
amster import.sh update - to remove wait for ds pods. Now that AM has a robust ready endpoint this is not needed.

refs: CLOUD-2839, CLOUD-2847