feat: Base images are based on nightly forgeops image builds
Moving the builds back over to forgeops-public. This is required because we need a secret to pull platform images repo which lives in forgeops-public and will leave builds isolated from engineering-devops.
I had to drop the usage of "./bin/cdk build" because it requires access
to a cluster in order to run build. There is no cluster in that project.

Additional drops rcs-agent and ui images from image-defaulter and
removes hard coded images in deployment objects in favor of
kustomization.yaml setting this.

ref: CLOUD-3241