RELENG-1321 Apply ForgeOps PIT#2 promotion to Platform Images repo
This is a workaround for ID Cloud branches 2021.4 through 2021.8 only.

In order to allow the FRaaS team time to migrate their scripts to
consume configuration from the Platform Images repo instead of the
ForgeOps repo, perform the PIT#2 promotions in both repos for ID Cloud
branches 2021.4 to 2021.8.

To avoid running the PIT#2 tests in both repos, we run the PIT#2 tests
using the ForgeOps pipeline only, then apply the promotion in both repos.
The Platform Images promotion process may be incomplete on ID Cloud
branches 2021.4 - 2021.8, so it's better to use the ForgeOps pipelines.