refactor: Remove the am-cdk and idm-cdk Kustomize bases
Align the behaviour in am and idm bases with am-cdk and idm-cdk to work with both ds-operator and legacy deployments.
Remove the am and idm base folders.
Rename the am-cdk and idm-cdk bases to am and idm.
Update the forgeops and utils scripts to work with the new base structure.

ref: FORGEOPS-4386