fix: Fix backup/restore backends and disaster recovery steps
Define backends to backup as not all backends are replicated and cause errors.
Update the the disaster recovery steps to not check the basedn for the schema backend.
Remove unrequired password change step in init-and-restore. This was added due to an error generated when backing up the password backends.
Remove duplicate env setting in env.sh
Comment out sample schema code.

ref: FORGEOPS-4959