fix: Fix memory values in t-shirt size overlays
After previously setting memory limits and requests to the same value as widely recommended, I set them both to the lower amount based on the memory request.  This could potentially limit the performance of the platform so this PR realigns the limits and requests to the previous memory limit value.

ref: FORGEOPS-5171