refactor: Generate random job names
Use needsHashSuffix annotation for amster and ds-set-passwords to generate random name suffix for Kustomize.
generateName doesn't work with Kustomize.
Increase TTL to allow for troubleshooting after deployment.
Don't delete amster in Helm after deployment to allow it to remain for the TTL.
Add force option to amster in Helm to allow redeploy.

ref: FORGEOPS-6279